Move various components into components package
